proto: enable reuse of mmaps between artifacts
This is a crude prototype which allows reusing the `CodeMemory` between
distinct Artifacts and Instances. This can reduce the overhead of
allocating mmaps with the kernel significantly. It turns out that the
kernel would still spend a lot of time if requested to change the
protection of a mapping from RX to RW.

This overhead is entirely mitigated by using HUGETLB, which is sadly
non-portable and is a major pain to enable (requires a kernel parameter).

With these two things in place the sys time goes down to almost zero
when loading contracts into memory, and the overall time of the
`deserialize/10000` benchamrk is reduced from ~2ms per iteration down to
1.25ms.