Do not allocate any registers to locals
This is a weird one. Before we'd alocate a couple arbitrary locals into
function-wide registers.

I suspect this was initially added in order to improve the throughput
for a manufactured micro-benchmark which had a loop using the first
couple locals to store the loop counter. In practice however, we have no
guarantee that the compiler will prefer low-numbered slots for locals it
expects to use frequently. Not to mention it doesn't actually help in
any way in the worst-case.

We don't actually reuse the registers that have been made free by this
change in any way so I would expect this change to be a net-negative in
terms of wasm execution performance. On the other hand, this would
probably improve the compilation performance slightly for evil
contracts that specify a large number of locals.